decouple NominatimConnector and NominatimUpdater class
The actual importer is now named NominatimImporter, while
NominatimConnector becomes the class for the common code shared
between the two classes.
